requirements.txt
altgraph==0.16.1
asgiref==3.2.3
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
cx-Freeze==5.1.1
Django==2.2.9
django-jenkins==0.110.0
djangorestframework==3.11
entrypoints==0.3
flake8==3.7.9
future==0.16.0
httpie==1.0.3
idna==2.7
lxml==4.3.1
macholib==1.11
mccabe==0.6.1
pefile==2018.8.8
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.2.0
PyInstaller==3.6
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.20.0
sqlparse==0.3.0
urllib3==1.24.2
virtualenv==16.0.0
